MCR-2398 MCRXMLMetadataManager Interface
Improve javadocs.
